+++ CHANGELOG.md
@@ -8,5 +8,6 @@ Created on Dec 20, 2023
 
 ## Changelog
 
-- Initial Create: initial created on December 14, 2023 -> **0.00**
+- Initial Create: Initial created on December 14, 2023 -> **0.00**
 - Initial Development: Development started on Decemeber 20, 2023 -> **0.1**
+- Initial Development: Authentication processes work on Decemeber 21, 2023 -> **0.2**